#192 Allow to skip the ssl verification
