+++ deploy/service.yaml
@@ -115,6 +115,8 @@ spec:
               value: 'true'
             - name: GENERATE_METRIC_EVENTS
               value: 'true'
+            - name: HTTP_SSL_VERIFY
+              value: 'true'
         - name: distributor
           image: keptn/distributor:0.7.1
           ports:
